break out metadata into model
