+++ plugins/amazonq/shared/jetbrains-community/src/software/aws/toolkits/jetbrains/services/amazonq/lsp/model/ExtendedClientMetadata.kt
@@ -1,0 +1,43 @@
+// Copyright 2025 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package software.aws.toolkits.jetbrains.services.amazonq.lsp.model
+
+import software.aws.toolkits.jetbrains.services.telemetry.ClientMetadata
+
+data class ExtendedClientMetadata(
+    val aws: AwsMetadata,
+)
+
+data class AwsMetadata(
+    val clientInfo: ClientInfoMetadata,
+)
+
+data class ClientInfoMetadata(
+    val extension: ExtensionMetadata,
+    val clientId: String,
+    val version: String,
+    val name: String,
+)
+
+data class ExtensionMetadata(
+    val name: String,
+    val version: String,
+)
+
+fun createExtendedClientMetadata(): ExtendedClientMetadata {
+    val metadata = ClientMetadata.getDefault()
+    return ExtendedClientMetadata(
+        aws = AwsMetadata(
+            clientInfo = ClientInfoMetadata(
+                extension = ExtensionMetadata(
+                    name = metadata.awsProduct.toString(),
+                    version = metadata.awsVersion
+                ),
+                clientId = metadata.clientId,
+                version = metadata.parentProductVersion,
+                name = metadata.parentProduct
+            )
+        )
+    )
+}
